improvement and new thing types are added
The binding is re-structured to be able to handle the multiple channel devices easier in the OpenHAB. The multiple channel devices are added in multiple times to the list with notification of the channel. e.g. a Sensor/Actuator 2/2 is added two times.

The scene tigger is implemented, where the switch is acting like a real tigger, after its activation the state is going back to off, to be able to activate again.

Current devices:
Actuator (simple one channel actuators are tested, multiple channel actuators are not tested(
Thermostat (tested)
Rule (tested)
Scene (tested)
Shutter actuator (not tested)
Window/Door position and open/close sensor
Door/ring sensor (not tested)
Dimming actuator (not tested)